backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.13.2",
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.121.3",
    "httpx>=0.28.1",
    "openai>=2.8.1",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "websockets>=15.0.1",
]

[tool.pytest.ini_options]
pythonpath = ["."]
